#322 - further refactoring and package consolidation.
