+++ src/main/java/com/homihq/db2rest/d1/D1SchemaManager.java
@@ -1,10 +1,11 @@
-package com.homihq.db2rest.schema;
+package com.homihq.db2rest.d1;
 
 import com.homihq.db2rest.d1.D1RestClient;
 import com.homihq.db2rest.d1.model.D1Column;
 import com.homihq.db2rest.d1.model.D1Table;
 import com.homihq.db2rest.dialect.Dialect;
 import com.homihq.db2rest.model.DbTable;
+import com.homihq.db2rest.schema.SchemaManager;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -14,7 +15,7 @@ import java.util.List;
 
 @Slf4j
 @RequiredArgsConstructor
-public class D1SchemaManager implements SchemaManager{
+public class D1SchemaManager implements SchemaManager {
 
     private final D1RestClient d1RestClient;
