+++ src/main/java/com/homihq/db2rest/jdbc/config/JdbcDataSourceConfiguration.java
@@ -1,6 +1,5 @@
-package com.homihq.db2rest.config;
+package com.homihq.db2rest.jdbc.config;
 
-import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -11,9 +10,9 @@ import org.springframework.context.annotation.Profile;
 import javax.sql.DataSource;
 
 @Configuration
-@Profile("!it")
+@Profile("!it") //integration test profile auto-configures DB with testcontainer
 @ConditionalOnProperty(prefix = "db2rest.datasource", name = "type", havingValue = "jdbc")
-public class JdbcConfiguration {
+public class JdbcDataSourceConfiguration {
 
     @Bean
     @ConfigurationProperties("spring.datasource")
@@ -29,4 +28,5 @@ public class JdbcConfiguration {
                 .build();
 
     }
+
 }
